+++ webpack.config.js
@@ -13,13 +13,11 @@ module.exports = {
     alias: {
       '@': path.resolve(__dirname, 'src/'),
       '@components': path.resolve(__dirname, 'src/components/'),
-      '@todoList': path.resolve(__dirname, 'src/components/TodoList/'),
-      '@userList': path.resolve(__dirname, 'src/components/UserList/'),
       '@constants': path.resolve(__dirname, 'src/constants/'),
-      '@core': path.resolve(__dirname, 'src/core/'),
       '@store': path.resolve(__dirname, 'src/store/'),
       '@utils': path.resolve(__dirname, 'src/utils/'),
       '@api': path.resolve(__dirname, 'src/api/'),
+      '@template': path.resolve(__dirname, 'src/template/'),
     },
     extensions: ['.js'],
   },
